+++ src/common.h
@@ -71,6 +71,12 @@ namespace sharp {
     int textSpacing;
     VipsTextWrap textWrap;
     int textAutofitDpi;
+    bool joinAnimated;
+    int joinAcross;
+    int joinShim;
+    std::vector<double> joinBackground;
+    VipsAlign joinHalign;
+    VipsAlign joinValign;
     std::vector<double> pdfBackground;
 
     InputDescriptor():
@@ -79,7 +85,7 @@ namespace sharp {
       failOn(VIPS_FAIL_ON_WARNING),
       limitInputPixels(0x3FFF * 0x3FFF),
       unlimited(false),
-      access(VIPS_ACCESS_RANDOM),
+      access(VIPS_ACCESS_SEQUENTIAL),
       bufferLength(0),
       isBuffer(false),
       density(72.0),
@@ -108,6 +114,12 @@ namespace sharp {
       textSpacing(0),
       textWrap(VIPS_TEXT_WRAP_WORD),
       textAutofitDpi(0),
+      joinAnimated(false),
+      joinAcross(1),
+      joinShim(0),
+      joinBackground{ 0.0, 0.0, 0.0, 255.0 },
+      joinHalign(VIPS_ALIGN_LOW),
+      joinValign(VIPS_ALIGN_LOW),
       pdfBackground{ 255.0, 255.0, 255.0, 255.0 } {}
   };
 
@@ -147,6 +159,7 @@ namespace sharp {
     FITS,
     EXR,
     JXL,
+    RAD,
     VIPS,
     RAW,
     UNKNOWN,
@@ -231,12 +244,6 @@ namespace sharp {
     Set embedded profile.
   */
   VImage SetProfile(VImage image, std::pair<char*, size_t> icc);
-
-  /*
-    Does this image have an alpha channel?
-    Uses colour space interpretation with number of channels to guess this.
-  */
-  bool HasAlpha(VImage image);
 
   /*
     Remove all EXIF-related image fields.
@@ -368,7 +375,7 @@ namespace sharp {
   std::tuple<VImage, std::vector<double>> ApplyAlpha(VImage image, std::vector<double> colour, bool premultiply);
 
   /*
-    Removes alpha channel, if any.
+    Removes alpha channels, if any.
   */
   VImage RemoveAlpha(VImage image);